@@ -47,7 +47,7 @@ class CoachMark {
   CoachMark({this.bgColor = const Color(0xB2212121)});
 
   /// Global key to get an access for CoachMark's State
-  GlobalKey<_HighlighterCoachMarkState> globalKey;
+  GlobalKey<_HighlighterCoachMarkState>? globalKey;
 
   /// Background color
   Color bgColor;
@@ -59,10 +59,10 @@ class CoachMark {
   bool get isVisible => _isVisible;
 
   /// Called when CoachMark is closed
-  VoidCallback _onClose;
+  VoidCallback? _onClose;
 
   /// Contains OverlayEntry with CoachMark's Widget
-  OverlayEntry _overlayEntryBackground;
+  OverlayEntry? _overlayEntryBackground;
 
   /// Brings out CoachMark's widget with animation on the whole screen
   ///
@@ -84,15 +84,15 @@ class CoachMark {
   ///
   /// Callback [onClose] is called when CoachMark is closed
   void show({
-    @required BuildContext targetContext,
-    @required List<Widget> children,
-    @required Rect markRect,
+    required BuildContext targetContext,
+    required List<Widget> children,
+    required Rect markRect,
     BoxShape markShape = BoxShape.circle,
-    Duration duration,
-    VoidCallback onClose,
-    HitTestBehavior hitTestBehavior,
-    bool closeOnTap,
-    VoidCallback onTransparentSectionTap,
+    Duration? duration,
+    VoidCallback? onClose,
+    HitTestBehavior? hitTestBehavior,
+    bool? closeOnTap,
+    VoidCallback? onTransparentSectionTap,
   }) async {
     // Prevent from showing multiple marks at the same time
     if (_isVisible) {
@@ -114,14 +114,14 @@ class CoachMark {
                 markShape: markShape,
                 doClose: close,
                 children: children,
-                hitTestBehavior: hitTestBehavior,
-                closeOnTap: closeOnTap,
+                hitTestBehavior: hitTestBehavior ?? HitTestBehavior.translucent,
+                closeOnTap: closeOnTap ?? false,
                 onTransparentSectionTap: onTransparentSectionTap,
               ),
         );
 
     OverlayState overlayState = Overlay.of(targetContext);
-    overlayState.insert(_overlayEntryBackground);
+    overlayState.insert(_overlayEntryBackground!);
 
     if (duration != null) {
       await new Future.delayed(duration).then((_) => close());
@@ -132,11 +132,11 @@ class CoachMark {
   Future close() async {
     if (_isVisible) {
       await globalKey?.currentState?.reverse();
-      _overlayEntryBackground.remove();
+      _overlayEntryBackground!.remove();
 
       _isVisible = false;
       if (_onClose != null) {
-        _onClose();
+        _onClose!();
       }
     }
   }
@@ -146,12 +146,12 @@ class CoachMark {
 /// [markRect]
 class _HighlighterCoachMarkWidget extends StatefulWidget {
   _HighlighterCoachMarkWidget({
-    Key key,
-    @required this.markRect,
-    @required this.markShape,
-    @required this.children,
-    @required this.doClose,
-    @required this.bgColor,
+    Key? key,
+    required this.markRect,
+    required this.markShape,
+    required this.children,
+    required this.doClose,
+    required this.bgColor,
     this.hitTestBehavior = HitTestBehavior.translucent,
     this.closeOnTap = true,
     this.onTransparentSectionTap,
@@ -164,7 +164,7 @@ class _HighlighterCoachMarkWidget extends StatefulWidget {
   final Color bgColor;
   final HitTestBehavior hitTestBehavior;
   final bool closeOnTap;
-  final VoidCallback onTransparentSectionTap;
+  final VoidCallback? onTransparentSectionTap;
 
   @override
   _HighlighterCoachMarkState createState() => new _HighlighterCoachMarkState();
@@ -172,9 +172,9 @@ class _HighlighterCoachMarkWidget extends StatefulWidget {
 
 class _HighlighterCoachMarkState extends State<_HighlighterCoachMarkWidget>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<double> _blurAnimation;
-  Animation<double> _opacityAnimation;
+  late AnimationController _controller;
+  late Animation<double> _blurAnimation;
+  late Animation<double> _opacityAnimation;
 
   //Does reverse animation, called when coachMark is closing.
   Future reverse() {
@@ -221,10 +221,9 @@ class _HighlighterCoachMarkState extends State<_HighlighterCoachMarkWidget>
   Widget build(BuildContext context) {
     Rect position = widget.markRect;
     final clipper = _CoachMarkClipper(position);
-
     return AnimatedBuilder(
         animation: _controller,
-        builder: (BuildContext context, Widget child) {
+        builder: (BuildContext context, Widget? child) {
           return GestureDetector(
             behavior: widget.hitTestBehavior,
             onTap: () {
@@ -232,7 +231,7 @@ class _HighlighterCoachMarkState extends State<_HighlighterCoachMarkWidget>
                 _onPointer(null);
               }
               if (widget.onTransparentSectionTap != null) {
-                widget.onTransparentSectionTap();
+                widget.onTransparentSectionTap!();
               }
             },
             child: Stack(
@@ -289,7 +288,7 @@ class _HighlighterCoachMarkState extends State<_HighlighterCoachMarkWidget>
         });
   }
 
-  void _onPointer(PointerEvent p) {
+  void _onPointer(PointerEvent? p) {
     widget.doClose();
   }
 }
@@ -298,14 +297,14 @@ class _HighlighterCoachMarkState extends State<_HighlighterCoachMarkWidget>
 /// overrides a special hitTest
 class _CoachMarkLayer extends Listener {
   const _CoachMarkLayer(
-      {Key key,
+      {Key? key,
       onPointerDown,
       onPointerMove,
       onPointerUp,
       onPointerCancel,
       behavior,
-      this.markPosition,
-      Widget child})
+      required this.markPosition,
+      Widget? child})
       : super(
             key: key,
             onPointerDown: onPointerDown,
@@ -350,9 +349,9 @@ class _RenderPointerListenerWithExceptRegion extends RenderPointerListener {
       onPointerMove,
       onPointerUp,
       onPointerCancel,
-      HitTestBehavior behavior,
-      this.exceptRegion,
-      RenderBox child})
+      required HitTestBehavior behavior,
+      required this.exceptRegion,
+      RenderBox? child})
       : super(
             onPointerDown: onPointerDown,
             onPointerMove: onPointerMove,
@@ -364,13 +363,13 @@ class _RenderPointerListenerWithExceptRegion extends RenderPointerListener {
   final Rect exceptRegion;
 
   @override
-  bool hitTest(HitTestResult result, {Offset position}) {
+  bool hitTest(BoxHitTestResult result, {Offset? position}) {
     bool hitTarget = false;
-    if (exceptRegion.contains(position)) {
+    if (position != null && exceptRegion.contains(position)) {
       result.add(new BoxHitTestEntry(this, position));
       return false;
     }
-    if (size.contains(position)) {
+    if (position != null && size.contains(position)) {
       hitTarget =
           hitTestChildren(result, position: position) || hitTestSelf(position);
       if (hitTarget || behavior == HitTestBehavior.translucent)
@@ -398,9 +397,9 @@ class _CoachMarkClipper extends CustomClipper<Path> {
 ///This class makes edges of hole blurred.
 class _CoachMarkPainter extends CustomPainter {
   _CoachMarkPainter({
-    @required this.rect,
-    @required this.shadow,
-    this.clipper,
+    required this.rect,
+    required this.shadow,
+    required this.clipper,
     this.coachMarkShape = BoxShape.circle,
   });
 
