@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
 
-import 'package:flutter_range_slider/flutter_range_slider.dart' as rngslider;
-
 class SkillsShowcase extends StatelessWidget {
   static final GlobalKey sliderKey = GlobalObjectKey("sliderKey");
 
@@ -21,50 +19,13 @@ class SkillsShowcase extends StatelessWidget {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: <Widget>[]
                     ..add(Text("Communication",
-                        style: textTheme.body1
-                            .copyWith(color: Colors.white70, fontSize: 16.0)))
-                    ..add(
-                      rngslider.RangeSlider(
-                        min: 0.0,
-                        max: 100.0,
-                        lowerValue: 0.0,
-                        upperValue: 70.0,
-                        divisions: 10,
-                        showValueIndicator: true,
-                        valueIndicatorMaxDecimals: 1,
-                        onChanged:
-                            (double newLowerValue, double newUpperValue) {},
-                      ),
-                    )))
+                        style: TextStyle(color: Colors.white70, fontSize: 16.0)))
+                    ))
               ..add(Text("Decision Making",
-                  style: textTheme.body1
-                      .copyWith(color: Colors.white70, fontSize: 16.0)))
-              ..add(
-                rngslider.RangeSlider(
-                  min: 0.0,
-                  max: 100.0,
-                  lowerValue: 0.0,
-                  upperValue: 90.0,
-                  divisions: 10,
-                  showValueIndicator: true,
-                  valueIndicatorMaxDecimals: 1,
-                  onChanged: (double newLowerValue, double newUpperValue) {},
-                ),
-              )
+                  style: TextStyle(color: Colors.white70, fontSize: 16.0)))
+
               ..add(Text("Leadership",
-                  style: textTheme.body1
-                      .copyWith(color: Colors.white70, fontSize: 16.0)))
-              ..add(
-                rngslider.RangeSlider(
-                  min: 0.0,
-                  max: 100.0,
-                  lowerValue: 0.0,
-                  upperValue: 30.0,
-                  divisions: 10,
-                  showValueIndicator: true,
-                  valueIndicatorMaxDecimals: 1,
-                  onChanged: (double newLowerValue, double newUpperValue) {},
-                ),
-              )));
+                  style: TextStyle(color: Colors.white70, fontSize: 16.0)))
+              ));
   }
 }
